+++ changelog.md
@@ -1,5 +1,6 @@
 ### Next version
 * InMemoryEventStore now checks for duplicate events. You can no longer write two events with the same cloud event id and source to the same stream.
+* Fixed an issue with command composition in Kotlin in which, in version 0.14.5, returned _all_ events in a stream and not only _new_ events. 
 
 ### 0.14.5 (2022-07-29)
 * Updated Kotlin extension functions for partial function application (`org.occurrent.application.composition.command.PartialExtensions`)